+++ include/Fighter.h
@@ -2,20 +2,19 @@
 #include "AnimatedObject.h"
 #include "LiveObject.h"
 #include "Castle.h"
-#include "Animation.h"
 
 class Fighter : public AnimatedObject, public LiveObject {
 public:
 	Fighter() = default ;
 	Fighter(const sf::Vector2f&,
-		const bool&, const int&, const float&, const float&);
+		const bool&, const int&, const int&, const int&);
 
 	void addLevel();
-	const float& getLevel() const;
-	const float& getDefaultAttack() const;
+	const int& getLevel() const;
+	const int& getDefaultAttack() const;
 	void setFullyDead() const;
 	virtual std::shared_ptr<bool> fullyDead() const;
-	float getGoldWorth() const;
+	int getGoldWorth() const;
 
 	virtual void attack(const std::shared_ptr<Fighter>& firstEnemy,
 		const std::shared_ptr<Castle>& Castle) = 0;
@@ -26,9 +25,9 @@ public:
 	virtual sf::Sprite create(float) const = 0;
 
 private:
-	float m_level;
-	float m_defaultGoldWorth;
-	float m_defaultAttack;
+	int m_level;
+	int m_defaultGoldWorth;
+	int m_defaultAttack;
 	float m_lastMovement;
 	//float m_movementSpeed; //  = milliseconds per one_pixel
 	//float m_attackSpeed;   //  = milliseconds per one_attack
